+++ src/material-table.js
@@ -384,6 +384,8 @@ export default class MaterialTable extends React.Component {
           tooltip: localization.bulkEditTooltip,
           position: 'toolbar',
           hidden: this.dataManager.bulkEditOpen,
+          disabled:
+            calculatedProps.isBulkEditable && calculatedProps.isBulkEditable(),
           onClick: () => {
             this.dataManager.changeBulkEditOpen(true);
             this.props.onBulkEditOpen && this.props.onBulkEditOpen(true);
@@ -898,7 +900,7 @@ export default class MaterialTable extends React.Component {
       return (
         <Table>
           <TableFooter style={{ display: 'grid' }}>
-            <TableRow>
+            <TableRow style={{ display: 'grid' }}>
               <props.components.Pagination
                 classes={{
                   root: props.classes.paginationRoot,